Rewrite nrf27 mini lib

+++ Firmware/Arduino/src/util.h
@@ -1,7 +1,7 @@
 #ifndef UTIL_H
 #define UTIL_H
 
-extern void array_set_bit(uint8_t* buf, const int off, const uint8_t val);
+extern void array_set_bit(uint8_t* buf, const uint16_t off, const uint8_t val);
 
 extern const int16_t map(const int16_t x, const int16_t in_min, const int16_t in_max, const int16_t out_min, const int16_t out_max);
